requirements.txt
# Core dependencies - minimal set for core functionality
click>=8.1.0
rich>=13.7.0
loguru>=0.7.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
google-generativeai>=0.7.2
httpx>=0.27.0
playwright>=1.48.0
scikit-learn>=1.5.0

# Desktop automation
pyautogui>=0.9.54
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0